Add SlackTranslator concrete Translator and the Supervisor
I am still debugging some AWS access issues with this version of the
code, however, a lot of this is working and starting to come together
so this is a checkpoint commit

AWAT now starts up, queries the database for existing work, and begins
processing it automatically. It can also receive requests for work and
track those requests at this time.

The CLI client is also fully functioning for get/list/start
Translations as well, at this time.