This got reverted during testing incorrectly
The tests were running against an old version of Mattermost, which is
why this bug reappeared. Older versions of Mattermost prepend this path,
and the newer ones don't, so we /do/ have to prepend it here.

I removed the prepended directory as part of 3f27f18 because of
accidental testing against an old version of Mattermost, where this
line will cause a double-prepend. This change reverts that portion of
3f27f18 so that /data is prepended, as it needs to be.